pyproject.toml
[project]
name = "main"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "dagger-io",
    "qrcode>=8.0",
    "pillow>=11.0.0",
]

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"
